Task
Summer weather can be quite unpredictable, and Victor needs your assistance. Write a program that, based on the time of day and the temperature, recommends to Victor what clothes to wear. Your friend has different plans for each stage of the day that require varied clothing, as indicated in the table.
Two lines are read from the console:
•	Temperature - an integer in the range [10... 42]
•	Text, time of day - with possibilities - "Morning", "Afternoon", "Evening"