Task
Martin and Nelly are buying a new house not far from Sofia. Nelly loves flowers so much that she convinces you to write a program to calculate how much it will cost them, to plant a certain number of flowers and whether the available budget will be enough for them. Different flowers come at different prices.