Task
Write a program that reads two six-digit integers from the console in the range from 100000 to 300000. The first entered number will always be smaller than the second one. Print on one line, separated by spaces, all numbers that are between the two numbers red from the console and satisfy the following condition:
•	The sum of the digits at even and odd positions must be equal. If there are no numbers that meet this condition, do not output any result on the console.