Create DrumSet
Test a given method which takes in a decimal representing money, list of integers representing quality of drums and a list of strings representing commands. 
For each command each drum is hit by the number indicated. If a drum breaks (goes below or equal to zero) it must be replaced by the initial amount, then initial amount times three must be reduced from the money.  If a new drum can’t be afforded it is removed from the list.
When the command "Hit it again, Gabsy!" is received, the method ends, and returns the current state of the drum set alongside the leftover money.